improve greatest eclipse and greatest duration circumstances
